requirements.txt
boto3
browser-use
langchain-google-genai
playwright
requests
asyncio-timeout
python-dotenv